requirements.txt
cached_property==1.5.2
camelot_py==0.10.1
Flask==2.1.2
numpy==1.21.6
pandas==1.3.5
Werkzeug==2.1.2
xlsxwriter==3.0.3
opencv-python